Refactor: @/component/common -> @component로 import 방식 변경
