+++ src/components/Survey/SplitType/SplitType.tsx
@@ -5,7 +5,7 @@ import Workout from '@/components/Survey/Workout/Workout';
 import FullBodyWorkout from '@/assets/svg/full-body-workout.svg';
 import ThreeBodyWorkout from '@/assets/svg/three-body-workout.svg';
 import FiveBodyWorkout from '@/assets/svg/five-body-workout.svg';
-import { Carousel, CustomButton } from '@/components/common';
+import { Carousel, CustomButton } from '@/components';
 
 const { s_container, s_title, s_highlight, s_spiltCarousel, s_noProgressBarTitle } = style;
