+++ src/components/Survey/SurveyComplete/SurveyComplete.tsx
@@ -8,7 +8,7 @@ import { SurveyFields } from '@/types';
 import { usePatchUserInfo } from '@/hooks/api';
 import { HOME_ACTIVITY } from '@/consts';
 import { startActivity } from '@/utils/mobile/token';
-import { CustomButton } from '@/components/common';
+import { CustomButton } from '@/components';
 
 const { s_container, s_lottieContainer, s_content } = style;
